making the r0 changes more stable and visually cohesive
